requirements.txt
requests
openai
itchat-uos
pyinstaller
beautifulsoup4
pydantic
PyYAML
sqlalchemy

# git+https://github.com/devcxl/wechat-api.git